+++ outbound-dialer/src/main/java/org/motechproject/nms/outbounddialer/repository/CallDetailRecordDataService.java
@@ -1,0 +1,7 @@
+package org.motechproject.nms.outbounddialer.repository;
+
+import org.motechproject.mds.service.MotechDataService;
+import org.motechproject.nms.outbounddialer.domain.CallDetailRecord;
+
+public interface CallDetailRecordDataService extends MotechDataService<CallDetailRecord> {
+}
