--- outbound-dialer/src/test/resources/META-INF/spring/test.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xmlns:osgi="http://www.eclipse.org/gemini/blueprint/schema/blueprint"
-       xsi:schemaLocation="http://www.springframework.org/schema/beans
-    http://www.springframework.org/schema/beans/spring-beans-3.1.xsd
-    http://www.eclipse.org/gemini/blueprint/schema/blueprint
-    http://www.eclipse.org/gemini/blueprint/schema/blueprint/gemini-blueprint.xsd">
-
-    <import resource="classpath*:META-INF/motech/*.xml" />
-
-    <osgi:reference id="settingsService"
-                    interface="org.motechproject.nms.outbounddialer.service.SettingsService" />
-
-</beans>
