+++ outbound-dialer/src/test/resources/test-files/cdrSummary_OBD_20150506070809.csv
@@ -1,0 +1,3 @@
+ece70579-c756-4f03-af42-a9631ba76589,imi123,1111111111,,0,,w1_1.wav,1,c123,aa,1,1001,1
+42a7f6bc-68e7-4c28-a9f4-ed3993c61676,imi123,2222222222,,0,,w2_1.wav,2,c456,bb,3,3001,1
+41e35bbd-62ef-4e72-afb1-284b0230d73c,imi123,3333333333,,0,,w3_2.wav,3,c789,cc,2,2003,4
